+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Added mod sha256 digest verification
 
+### Fixed
+
+- Fixed game extensions not being handled
+
 ## [0.9.9] 2024-03-17
 
 ### Changed